Bluetooth: Only check SAR bits if frame is an I-frame
The SAR bits doesn't make sense for an S-frame. It doesn't use SAR.

Checking SAR for a S-frames can lead to L2CAP errors, it could close
the channel with an invalid packet length, since we was removing the 2
of the of any frame that match SAR start bits, without check if it is
an I-frame.

Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit bc1b1f8bee63966649dd5ac7d10d31a6556bf19b)

Change-Id: Id5f2527020266b8f078a8f77881023d404d79116
Signed-off-by: Mat Martineau <mathewm@codeaurora.org>
Signed-off-by: Andika Edo <dhiika@phiexz.com>