mmc: core: Fix possible race between runtime suspend and runtime resume
There is a possible race with mmc_claim_host() in mmc_sd_suspend()/
mmc_suspend() and mmc_claim_host() in mmc_blk_issue_rq() when runtime
pm is enabled. It can happen only if a block request comes while
runtime suspend is in progress and exactly before calling
mmc_sd_suspend(). Since deselect of SD and MMC cards is not required
during suspend, it is removed as a workaround in this patch.

CRs-fixed: 274217
Change-Id: Ia311986c11641c2d8183d29b5d11817d401e4f03
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>
Signed-off-by: Andika Edo <dhiika@phiexz.com>